pyproject.toml
[tool.poetry]
name = "adminka-stats"
version = "0.1.0"
description = "Crypto Analytics Admin Panel"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.32.2"
pandas = "^2.2.0"
plotly = "^5.19.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.1"
streamlit-extras = "^0.4.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"